+++ qt/manageprofiles/__init__.py
@@ -41,6 +41,7 @@ import qttools
 import messagebox
 from manageprofiles.tab_general import GeneralTab
 from manageprofiles.tab_options import OptionsTab
+from manageprofiles.tab_expert_options import ExpertOptionsTab
 from editusercallback import EditUserCallback
 from restoreconfigdialog import RestoreConfigDialog
 
@@ -385,306 +386,8 @@ class SettingsDialog(QDialog):
         scrollArea = QScrollArea(self)
         scrollArea.setFrameStyle(QFrame.Shape.NoFrame)
         self.tabs.addTab(scrollArea, _('E&xpert Options'))
-
-        layoutWidget = QWidget(self)
-        layout = QVBoxLayout(layoutWidget)
-
-        label = QLabel(_('Caution: Change these options only if you really '
-                         'know what you are doing.'), self)
-        qttools.setFontBold(label)
-        layout.addWidget(label)
-
-        label = QLabel(_("Run 'rsync' with '{cmd}':").format(cmd='nice'))
-        layout.addWidget(label)
-        grid = QGridLayout()
-        grid.setColumnMinimumWidth(0, 20)
-        layout.addLayout(grid)
-
-        self.cbNiceOnCron = QCheckBox(
-            _('as cron job') + self.printDefault(
-                self.config.DEFAULT_RUN_NICE_FROM_CRON), self)
-        grid.addWidget(self.cbNiceOnCron, 0, 1)
-
-        self.cbNiceOnRemote = QCheckBox(
-            _('on remote host') + self.printDefault(
-                self.config.DEFAULT_RUN_NICE_ON_REMOTE), self)
-        grid.addWidget(self.cbNiceOnRemote, 1, 1)
-
-        label = QLabel(_("Run 'rsync' with '{cmd}':").format(cmd='ionice'))
-        layout.addWidget(label)
-        grid = QGridLayout()
-        grid.setColumnMinimumWidth(0, 20)
-        layout.addLayout(grid)
-
-        self.cbIoniceOnCron = QCheckBox(
-            _('as cron job') + self.printDefault(
-                self.config.DEFAULT_RUN_IONICE_FROM_CRON), self)
-        grid.addWidget(self.cbIoniceOnCron, 0, 1)
-
-        self.cbIoniceOnUser = QCheckBox(
-            _('when taking a manual snapshot') + self.printDefault(
-                self.config.DEFAULT_RUN_IONICE_FROM_USER), self)
-        grid.addWidget(self.cbIoniceOnUser, 1, 1)
-
-        self.cbIoniceOnRemote = QCheckBox(
-            _('on remote host') + self.printDefault(
-                self.config.DEFAULT_RUN_IONICE_ON_REMOTE), self)
-        grid.addWidget(self.cbIoniceOnRemote, 2, 1)
-
-        self.nocacheAvailable = tools.checkCommand('nocache')
-        txt = _("Run 'rsync' with '{cmd}':").format(cmd='nocache')
-
-        if not self.nocacheAvailable:
-            txt += ' ' + _("(Please install 'nocache' to enable this option)")
-        layout.addWidget(QLabel(txt))
-        grid = QGridLayout()
-        grid.setColumnMinimumWidth(0, 20)
-        layout.addLayout(grid)
-
-        self.cbNocacheOnLocal = QCheckBox(
-            _('on local machine') + self.printDefault(
-                self.config.DEFAULT_RUN_NOCACHE_ON_LOCAL), self)
-        self.cbNocacheOnLocal.setEnabled(self.nocacheAvailable)
-        grid.addWidget(self.cbNocacheOnLocal, 0, 1)
-
-        self.cbNocacheOnRemote = QCheckBox(
-            _('on remote host') + self.printDefault(
-                self.config.DEFAULT_RUN_NOCACHE_ON_REMOTE), self)
-        grid.addWidget(self.cbNocacheOnRemote, 2, 1)
-
-        self.cbRedirectStdoutInCron = QCheckBox(
-            _('Redirect stdout to /dev/null in cronjobs.')
-            + self.printDefault(self.config.DEFAULT_REDIRECT_STDOUT_IN_CRON),
-            self)
-        qttools.set_wrapped_tooltip(
-            self.cbRedirectStdoutInCron,
-            _('Cron will automatically send an email with attached output '
-              'of cronjobs if an MTA is installed.')
-        )
-        layout.addWidget(self.cbRedirectStdoutInCron)
-
-        self.cbRedirectStderrInCron = QCheckBox(
-            _('Redirect stderr to /dev/null in cronjobs.')
-            + self.printDefault(self.config.DEFAULT_REDIRECT_STDERR_IN_CRON),
-            self)
-        qttools.set_wrapped_tooltip(
-            self.cbRedirectStderrInCron,
-            _('Cron will automatically send an email with attached errors '
-              'of cronjobs if an MTA is installed.')
-        )
-        layout.addWidget(self.cbRedirectStderrInCron)
-
-        # bwlimit
-        hlayout = QHBoxLayout()
-        layout.addLayout(hlayout)
-        self.cbBwlimit = QCheckBox(_('Limit rsync bandwidth usage:'), self)
-        hlayout.addWidget(self.cbBwlimit)
-        self.spbBwlimit = QSpinBox(self)
-        self.spbBwlimit.setSuffix(' ' + _('KB/sec'))
-        self.spbBwlimit.setSingleStep(100)
-        self.spbBwlimit.setRange(0, 1000000)
-        hlayout.addWidget(self.spbBwlimit)
-        hlayout.addStretch()
-        enabled = lambda state: self.spbBwlimit.setEnabled(state)
-        enabled(False)
-        self.cbBwlimit.stateChanged.connect(enabled)
-        qttools.set_wrapped_tooltip(
-            self.cbBwlimit,
-            [
-                "Uses 'rsync --bwlimit=RATE'. From 'man rsync':",
-                'This option allows you to specify the maximum transfer rate '
-                'for the data sent over the socket, specified in units per '
-                'second. The RATE value can be suffixed with a string to '
-                'indicate a size multiplier, and may be a fractional value '
-                '(e.g. "--bwlimit=1.5m").',
-                'If no suffix is specified, the value will be assumed to be '
-                'in units of 1024 bytes (as if "K" or "KiB" had been '
-                'appended).',
-                'See the --max-size option for a description of all the '
-                'available suffixes. A value of zero specifies no limit.'
-                '',
-                'For backward-compatibility reasons, the rate limit will be '
-                'rounded to the nearest KiB unit, so no rate smaller than '
-                '1024 bytes per second is possible.',
-                '',
-                'Rsync writes data over the socket in blocks, and this option '
-                'both limits the size of the blocks that rsync writes, and '
-                'tries to keep the average transfer rate at the requested '
-                'limit. Some "burstiness" may be seen where rsync writes out '
-                'a block of data and then sleeps to bring the average rate '
-                'into compliance.',
-                '',
-                'Due to the internal buffering of data, the --progress '
-                'option may not be an accurate reflection on how fast the '
-                'data is being sent. This is because some files can show up '
-                'as being rapidly sent when the data is quickly buffered, '
-                'while other can show up as very slow when the flushing of '
-                'the output buffer occurs. This may be fixed in a future '
-                'version.'
-            ]
-        )
-
-        self.cbPreserveAcl = QCheckBox(_('Preserve ACL'), self)
-        qttools.set_wrapped_tooltip(
-            self.cbPreserveAcl,
-            [
-                "Uses 'rsync -A'. From 'man rsync':",
-                'This option causes rsync to update the destination ACLs to '
-                'be the same as the source ACLs. The option also implies '
-                '--perms.',
-                '',
-                'The source and destination systems must have compatible ACL '
-                'entries for this option to work properly. See the '
-                '--fake-super option for a way to backup and restore ACLs '
-                'that are not compatible.'
-            ]
-        )
-        layout.addWidget(self.cbPreserveAcl)
-
-        self.cbPreserveXattr = QCheckBox(
-            _('Preserve extended attributes (xattr)'), self)
-        qttools.set_wrapped_tooltip(
-            self.cbPreserveXattr,
-            [
-                "Uses 'rsync -X'. From 'man rsync':",
-                'This option causes rsync to update the destination extended '
-                'attributes to be the same as the source ones.',
-                '',
-                'For systems that support extended-attribute namespaces, a '
-                'copy being done by a super-user copies all namespaces '
-                'except system.*. A normal user only copies the user.* '
-                'namespace. To be able to backup and restore non-user '
-                'namespaces as a normal user, see the --fake-super option.',
-                '',
-                'Note that this option does not copy rsyncs special xattr '
-                'values (e.g. those used by --fake-super) unless you repeat '
-                'the option (e.g. -XX). This "copy all xattrs" mode cannot be '
-                'used with --fake-super.'
-            ]
-        )
-        layout.addWidget(self.cbPreserveXattr)
-
-        self.cbCopyUnsafeLinks = QCheckBox(
-            _('Copy unsafe links (works only with absolute links)'), self)
-        qttools.set_wrapped_tooltip(
-            self.cbCopyUnsafeLinks,
-            [
-                "Uses 'rsync --copy-unsafe-links'. From 'man rsync':",
-                'This tells rsync to copy the referent of symbolic links that '
-                'point outside the copied tree. Absolute symlinks are also '
-                'treated like ordinary files, and so are any symlinks in the '
-                'source path itself when --relative is used. This option has '
-                'no additional effect if --copy-links was also specified.'
-            ]
-        )
-        layout.addWidget(self.cbCopyUnsafeLinks)
-
-        self.cbCopyLinks = QCheckBox(
-            _('Copy links (dereference symbolic links)'), self)
-        qttools.set_wrapped_tooltip(
-            self.cbCopyLinks,
-            [
-                "Uses 'rsync --copy-links'. From 'man rsync':",
-                'When symlinks are encountered, the item that they point to '
-                '(the referent) is copied, rather than the symlink. In older '
-                'versions of rsync, this option also had the side-effect of '
-                'telling the receiving side to follow symlinks, such as '
-                'symlinks to directories. In a modern rsync such as this one,'
-                ' you will need to specify --keep-dirlinks (-K) to get this '
-                'extra behavior. The only exception is when sending files to '
-                'an rsync that is too old to understand -K -- in that case, '
-                'the -L option will still have the side-effect of -K on that '
-                'older receiving rsync.'
-            ]
-        )
-        layout.addWidget(self.cbCopyLinks)
-
-        # one file system option
-        self.cbOneFileSystem = QCheckBox(
-            _('Restrict to one file system'), self)
-        qttools.set_wrapped_tooltip(
-            self.cbOneFileSystem,
-            [
-                "Uses 'rsync --one-file-system'. From 'man rsync':",
-                'This tells rsync to avoid crossing a filesystem boundary '
-                'when recursing. This does not limit the user\'s ability '
-                'to specify items to copy from multiple filesystems, just '
-                'rsync\'s recursion through the hierarchy of each directory '
-                'that the user specified, and also the analogous recursion '
-                'on the receiving side during deletion. Also keep in mind '
-                'that rsync treats a "bind" mount to the same device as '
-                'being on the same filesystem.'
-            ]
-        )
-        layout.addWidget(self.cbOneFileSystem)
-
-        # additional rsync options
-        hlayout = QHBoxLayout()
-        layout.addLayout(hlayout)
-        tooltip = _('Options must be quoted e.g. {example}.').format(
-            example='--exclude-from="/path/to/my exclude file"')
-        self.cbRsyncOptions = QCheckBox(
-            _('Paste additional options to rsync'), self)
-        self.cbRsyncOptions.setToolTip(tooltip)
-        hlayout.addWidget(self.cbRsyncOptions)
-        self.txtRsyncOptions = QLineEdit(self)
-        self.txtRsyncOptions.editingFinished.connect(
-            self._slot_rsync_options_editing_finished)
-        self.txtRsyncOptions.setToolTip(tooltip)
-        hlayout.addWidget(self.txtRsyncOptions)
-
-        enabled = lambda state: self.txtRsyncOptions.setEnabled(state)
-        enabled(False)
-        self.cbRsyncOptions.stateChanged.connect(enabled)
-
-        # ssh prefix
-        hlayout = QHBoxLayout()
-        layout.addLayout(hlayout)
-        self.cbSshPrefix = QCheckBox(_('Add prefix to SSH commands'), self)
-        tooltip = [
-            _('Prefix to run before every command on remote host.'),
-            _("Variables need to be escaped with \\$FOO. This doesn't touch "
-              'rsync. So to add a prefix for rsync use "{example_value}" with '
-              '{rsync_options_value}.').format(
-                  example_value=self.cbRsyncOptions.text(),
-                  rsync_options_value \
-                      ='--rsync-path="FOO=bar:\\$FOO /usr/bin/rsync"'),
-            '',
-            '{default}: {def_value}'.format(
-                default=_('default'),
-                def_value=self.config.DEFAULT_SSH_PREFIX)
-        ]
-        qttools.set_wrapped_tooltip(self.cbSshPrefix, tooltip)
-        hlayout.addWidget(self.cbSshPrefix)
-        self.txtSshPrefix = QLineEdit(self)
-        qttools.set_wrapped_tooltip(self.txtSshPrefix, tooltip)
-        hlayout.addWidget(self.txtSshPrefix)
-
-        enabled = lambda state: self.txtSshPrefix.setEnabled(state)
-        enabled(False)
-        self.cbSshPrefix.stateChanged.connect(enabled)
-
-        qttools.equalIndent(self.cbRsyncOptions, self.cbSshPrefix)
-
-        self.cbSshCheckPing = QCheckBox(_('Check if remote host is online'))
-        qttools.set_wrapped_tooltip(
-            self.cbSshCheckPing,
-            _('Warning: If disabled and the remote host is not available, '
-              'this could lead to some weird errors.')
-        )
-        self.cbSshCheckCommands = QCheckBox(
-            _('Check if remote host supports all necessary commands.'))
-        qttools.set_wrapped_tooltip(
-            self.cbSshCheckCommands,
-            _('Warning: If disabled and the remote host does not support all '
-              'necessary commands, this could lead to some weird errors.')
-        )
-        layout.addWidget(self.cbSshCheckPing)
-        layout.addWidget(self.cbSshCheckCommands)
-
-        #
-        layout.addStretch()
-        scrollArea.setWidget(layoutWidget)
+        self._tab_expert_options = ExpertOptionsTab(self)
+        scrollArea.setWidget(self._tab_expert_options)
         scrollArea.setWidgetResizable(True)
 
         # buttons
@@ -710,29 +413,6 @@ class SettingsDialog(QDialog):
         self.resize(size)
 
         self.finished.connect(self._slot_finished)
-
-    def _slot_rsync_options_editing_finished(self):
-        """When editing the rsync options is finished warn and remove
-        --old-args option if present.
-        """
-        txt = self.txtRsyncOptions.text()
-
-        if '--old-args' in txt:
-            # No translation for this message because it is a rare case.
-            messagebox.warning(
-                text='Found rsync flag "--old-args". That flag will be removed'
-                ' from the options because it conflicts with the flag "-s" '
-                '(also known as "--secluded-args" or "--protected-args") which'
-                ' is used by Back In Time to force the "new form of argument '
-                'protection" in rsync.',
-                widget_to_center_on=self
-            )
-
-            # Don't leave two-blank spaces between other arguments
-            txt = txt.replace('--old-args ', '')
-            txt = txt.replace(' --old-args', '')
-            txt = txt.replace('--old-args', '')
-            self.txtRsyncOptions.setText(txt)
 
     def addProfile(self):
         ret_val = QInputDialog.getText(self, _('New profile'), str())
@@ -909,31 +589,7 @@ class SettingsDialog(QDialog):
         self._tab_options.load_values()
 
         # TAB: Expert Options
-        self.cbNiceOnCron.setChecked(self.config.niceOnCron())
-        self.cbIoniceOnCron.setChecked(self.config.ioniceOnCron())
-        self.cbIoniceOnUser.setChecked(self.config.ioniceOnUser())
-        self.cbNiceOnRemote.setChecked(self.config.niceOnRemote())
-        self.cbIoniceOnRemote.setChecked(self.config.ioniceOnRemote())
-        self.cbNocacheOnLocal.setChecked(
-            self.config.nocacheOnLocal() and self.nocacheAvailable)
-        self.cbNocacheOnRemote.setChecked(self.config.nocacheOnRemote())
-        self.cbRedirectStdoutInCron.setChecked(
-            self.config.redirectStdoutInCron())
-        self.cbRedirectStderrInCron.setChecked(
-            self.config.redirectStderrInCron())
-        self.cbBwlimit.setChecked(self.config.bwlimitEnabled())
-        self.spbBwlimit.setValue(self.config.bwlimit())
-        self.cbPreserveAcl.setChecked(self.config.preserveAcl())
-        self.cbPreserveXattr.setChecked(self.config.preserveXattr())
-        self.cbCopyUnsafeLinks.setChecked(self.config.copyUnsafeLinks())
-        self.cbCopyLinks.setChecked(self.config.copyLinks())
-        self.cbOneFileSystem.setChecked(self.config.oneFileSystem())
-        self.cbRsyncOptions.setChecked(self.config.rsyncOptionsEnabled())
-        self.txtRsyncOptions.setText(self.config.rsyncOptions())
-        self.cbSshPrefix.setChecked(self.config.sshPrefixEnabled())
-        self.txtSshPrefix.setText(self.config.sshPrefix())
-        self.cbSshCheckPing.setChecked(self.config.sshCheckPingHost())
-        self.cbSshCheckCommands.setChecked(self.config.sshCheckCommands())
+        self._tab_expert_options.load_values()
 
         # update
         self.updateRemoveOlder()
@@ -1011,30 +667,7 @@ class SettingsDialog(QDialog):
         self._tab_options.store_values()
 
         # expert options
-        self.config.setNiceOnCron(self.cbNiceOnCron.isChecked())
-        self.config.setIoniceOnCron(self.cbIoniceOnCron.isChecked())
-        self.config.setIoniceOnUser(self.cbIoniceOnUser.isChecked())
-        self.config.setNiceOnRemote(self.cbNiceOnRemote.isChecked())
-        self.config.setIoniceOnRemote(self.cbIoniceOnRemote.isChecked())
-        self.config.setNocacheOnLocal(self.cbNocacheOnLocal.isChecked())
-        self.config.setNocacheOnRemote(self.cbNocacheOnRemote.isChecked())
-        self.config.setRedirectStdoutInCron(
-            self.cbRedirectStdoutInCron.isChecked())
-        self.config.setRedirectStderrInCron(
-            self.cbRedirectStderrInCron.isChecked())
-        self.config.setBwlimit(self.cbBwlimit.isChecked(),
-                               self.spbBwlimit.value())
-        self.config.setPreserveAcl(self.cbPreserveAcl.isChecked())
-        self.config.setPreserveXattr(self.cbPreserveXattr.isChecked())
-        self.config.setCopyUnsafeLinks(self.cbCopyUnsafeLinks.isChecked())
-        self.config.setCopyLinks(self.cbCopyLinks.isChecked())
-        self.config.setOneFileSystem(self.cbOneFileSystem.isChecked())
-        self.config.setRsyncOptions(self.cbRsyncOptions.isChecked(),
-                                    self.txtRsyncOptions.text())
-        self.config.setSshPrefix(self.cbSshPrefix.isChecked(),
-                                 self.txtSshPrefix.text())
-        self.config.setSshCheckPingHost(self.cbSshCheckPing.isChecked())
-        self.config.setSshCheckCommands(self.cbSshCheckCommands.isChecked())
+        self._tab_expert_options.store_values()
 
         return True
 
@@ -1278,14 +911,9 @@ class SettingsDialog(QDialog):
 
         enabled = active_mode in ('ssh', 'ssh_encfs')
 
-        self.cbNiceOnRemote.setEnabled(enabled)
-        self.cbIoniceOnRemote.setEnabled(enabled)
-        self.cbNocacheOnRemote.setEnabled(enabled)
+        self._tab_expert_options.update_items_state(enabled)
+
         self.cbSmartRemoveRunRemoteInBackground.setVisible(enabled)
-        self.cbSshPrefix.setVisible(enabled)
-        self.txtSshPrefix.setVisible(enabled)
-        self.cbSshCheckPing.setVisible(enabled)
-        self.cbSshCheckCommands.setVisible(enabled)
 
     def updateExcludeItems(self):
         for index in range(self.listExclude.topLevelItemCount()):
@@ -1362,10 +990,6 @@ class SettingsDialog(QDialog):
     def excludeCustomSortOrder(self, *args):
         self.listExcludeSortLoop = self.customSortOrder(
             self.listExclude.header(), self.listExcludeSortLoop, *args)
-
-    def printDefault(self, value):
-        return ' ' + _('(default: {})').format(
-            _('enabled') if value else _('disabled'))
 
     def restoreConfig(self, *args):
         RestoreConfigDialog(self).exec()
